requirements.txt
async-lru==1.0.2
certifi==2020.6.20
chardet==3.0.4
idna==2.10
ini-parser==1.1.0
install==1.3.3
pyaes==1.6.1
Pyrogram==0.17.1
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
requests==2.24.0
six==1.15.0
TgCrypto==1.2.1
urllib3==1.25.10
